Handle optional arp probes in non-privileged mode
When running without root privilege, we can't do arp scanning.
The arp create_host_probe() function will return NULL in this case.
However, when marking this probe as optional, we do not want to see any
complaints about this probe.

So, in a case like this, create a dummy probe that will always be
skipped.

Signed-off-by: Olaf Kirch <okir@suse.com>