Get rid of an annoying error message in fm_topo_state_send_probe()
We were looking at the timeout of the subordinate probe. But as that probe
has not been scheduled yet, this returns -1; and we would try to set that
as our timeout.

We obviously have no need to know when the subordinate probe expires; if
something bad happens with it, we'll find out.

Signed-off-by: Olaf Kirch <okir@suse.com>